Add files via upload
Starbucks, an American company that is the largest coffeehouse chain in the world. Its headquarters are in Seattle, Washington. Starbucks started as a roaster and retailer of whole bean and ground coffee, tea, and spices with a single store in Seattle’s Pike Place Market in 1971. The company now operates more than 24,000 retail stores in 70 countries. (29 March 2021) Today Starbucks is the largest coffeehouse company in the world, with 32,938 retail locations as of the first quarter of 2021, followed distantly by coffee shop chains such as Dunkin Donuts with about 10,000 restaurants, Tim Hortons with 4,300 outlets, and Costa Coffee with nearly 1,700 stores worldwide. This dataset includes a record for every Starbucks or subsidiary store location currently in operation as of February 2017. Does this project discuss what city or country has the highest number of Starbucks stores per capita? What two Starbucks locations are the closest in proximity to one another? What location on Earth is farthest from a Starbucks? How has Starbucks expanded overseas? We will be using Starbucks worldwide database dataset to the dashboard, which is going to create interactive, real-time visualizations where users can combine data sources, add filters, and drill down into specific information.